requirements.txt
openai>=1.0.0
anthropic>=0.8.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.11.0
numpy>=1.21.0
Pillow>=9.0.0
reportlab>=3.6.0
PyYAML>=6.0.0
svglib>=1.2.1
ray>=2.0.0
click>=8.0.0
PyMuPDF>=1.23.0
opencv-python>=4.5.0
scikit-image>=0.21.0
pytest>=7.4.0
ezdxf>=0.17.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0